pyproject.toml
[project]
name = "resolve-assistant"
version = "0.1.5"
description = "AI-powered video editing assistant using Gemini + DaVinci Resolve"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = [
    "davinci-resolve", "ai-editing", "gemini", "video-editing", "mcp",
    "model-context-protocol", "timeline", "footage-analysis", "claude",
    "llm-tools", "video-production", "automatic-editing",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastmcp>=3.0.0",
    "google-genai>=1.64.0",
    "opentimelineio>=0.18.1",
    "otio-fcp-adapter>=1.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
resolve-assistant = "resolve_assistant:main"

[project.urls]
Homepage = "https://github.com/jenkinsm13/resolve-assistant"
Repository = "https://github.com/jenkinsm13/resolve-assistant"
Issues = "https://github.com/jenkinsm13/resolve-assistant/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
